feat: Learning Rate Finder
The learning rate finder implements a technique to easily estimate
a range of learning rates that should perform well given the
current model setup. When the model architecture or other
hyperparameters are changed, the learning rate finder can be run in
order to find a new suitable LR range.

The learning rate finder starts training the model at a very low LR,
increasing it every batch. The batch loss is plotted vs training
steps and a figure is created from which a suitable LR range can be
determined.

This technique was first introduced by Leslie Smith in
https://arxiv.org/abs/1506.01186.

Former-commit-id: 99721c56ad4d2e680a81c69afe90cad69c2aff11